Add support for lat/lon/radius and bbox to REST operator API
These four parameters are documented as supported on
https://www.transit.land/documentation/rest-api/agencies#operator-request-parameters,
but to my surprise, I got back a very unfiltered list of results when using the
`lat`, `lon`, and `radius` parameters. These work on an agency request, but it
would be great at the operator level as well.

It looks like the lower-level GraphQL interface supports these parameters, so
we just need to make sure they are passed through appropriately, much like the
agency request.